Moved Breadcrumbs class and removed internal namespace.
