+++ ts/Extensions/Breadcrumbs/BreadcrumbsDefaults.ts
@@ -19,6 +19,7 @@
  * */
 
 import type BreadcrumbsOptions from './BreadcrumbsOptions';
+import type { LangOptions } from '../../Core/Options';
 
 import { Palette } from '../../Core/Color/Palettes.js';
 
@@ -33,7 +34,7 @@ import { Palette } from '../../Core/Color/Palettes.js';
  *
  * @private
  */
-const lang = {
+const lang: Partial<LangOptions> = {
     /**
      * @since 10.0.0
      * @product  highcharts
