+++ ts/Extensions/Drilldown.ts
@@ -19,7 +19,10 @@ import type {
 } from '../Core/Renderer/AlignObject';
 import type AnimationOptions from '../Core/Animation/AnimationOptions';
 import type BBoxObject from '../Core/Renderer/BBoxObject';
-import type { BreadcrumbOptions } from './Breadcrumbs/BreadcrumbsOptions';
+import type {
+    BreadcrumbOptions,
+    BreadcrumbsOptions
+} from './Breadcrumbs/BreadcrumbsOptions';
 import type { ButtonRelativeToValue } from '../Maps/MapNavigationOptions';
 import type ColorType from '../Core/Color/ColorType';
 import type {
@@ -54,7 +57,7 @@ const { seriesTypes } = SeriesRegistry;
 import SVGRenderer from '../Core/Renderer/SVG/SVGRenderer.js';
 import Tick from '../Core/Axis/Tick.js';
 import U from '../Core/Utilities.js';
-import Breadcrumbs from './Breadcrumbs.js';
+import Breadcrumbs from './Breadcrumbs/Breadcrumbs.js';
 
 const {
     addEvent,
@@ -209,6 +212,7 @@ declare global {
             );
             allowPointDrilldown?: boolean;
             animation?: (boolean|Partial<AnimationOptions>);
+            breadcrumbs?: BreadcrumbsOptions;
             drillUpButton?: DrilldownDrillUpButtonOptions;
             series?: Array<SeriesTypeOptions>;
         }
