+++ ts/masters/modules/drilldown.src.ts
@@ -12,7 +12,7 @@
 'use strict';
 import Highcharts from '../../Core/Globals.js';
 import '../../Extensions/Drilldown.js';
-import Breadcrumbs from '../../Extensions/Breadcrumbs.js';
+import Breadcrumbs from '../../Extensions/Breadcrumbs/Breadcrumbs.js';
 const G: AnyRecord = Highcharts;
 G.Breadcrumbs = Breadcrumbs;
 Breadcrumbs.compose(G.Chart, G.defaultOptions);
