+++ ts/masters/modules/treemap.src.ts
@@ -11,7 +11,7 @@
 'use strict';
 import Highcharts from '../../Core/Globals.js';
 import '../../Series/Treemap/TreemapSeries.js';
-import Breadcrumbs from '../../Extensions/Breadcrumbs.js';
+import Breadcrumbs from '../../Extensions/Breadcrumbs/Breadcrumbs.js';
 const G: AnyRecord = Highcharts;
 G.Breadcrumbs = Breadcrumbs;
 Breadcrumbs.compose(G.Chart, G.defaultOptions);
